Add: Navbar implementation and pages
