--- navigation/mainContainer.js
@@ -1,8 +1,0 @@
-import * as React from "react";
-import {View} from "react-native";
-
-export default function MainContainer() {
-	return (
-		<View></View>
-	);
-}
